+++ include/elio/sync/object_cache.hpp
@@ -526,7 +526,7 @@ private:
             std::weak_ptr<internals>(state_), state_->cfg_.sweep_interval,
             sweep_cancel_.get_token());
 
-        auto handle = coro::detail::task_access::release(t);
+        auto handle = coro::detail::task_access::release(std::move(t));
         handle.promise().detached_ = true;
         handle.promise().detach_from_parent();
         coro::promise_base::set_current_frame(old_frame);